Vision Library for Mineral Detection
Imported DogeCV, which is an OpenCV library which conveniently supports FTC element detection. Wrote a mineral detection class that takes in a mineral color as a parameter to the constructor and has methods to return whether or not mineral is being seen by screen. Then made a test OpMode class that, like the Vuforia test class, tests out the accuracy of mineral detection with DogeCV.