Pseudocode for Basic Auto
Pseudocode (with some written code for Vuforia Navigation Target tracking) to eventually detect the gold mineral, move to the alliance depot, and then move to park in the crater. For testing purposes, this program is being written for the red side, as if the robot started on the side of the lander closer to the red alliance depot. Once this works, code will be written to apply for all 4 starting positions.