Introduced Omniwheel Slide Drive class
We're using essentially the same drivetrain for both the first and second robot. So, I made a drivetrain class that models an omni wheel slide drive and takes in three motors (left, right, middle) plus a gamepad and runs several methods that will be used in opmodes. Next step is to incorporate it in hardware class and tele-op class