Field Mapping (potentially usable down the road)
This is actually pretty cool: with the visual coordinate map Mich created of the field, we notice that lots of the game elements line up pretty nicely. From this, I've created the basic structure of two classes: Vector and FieldMap. Vector holds a position and FieldMap maps each important game element to its position on the field. When paired with an effective position tracking system, the possibilities are endless.