Integrating while loop breaking to prevent stuck in stop()
Robot1_Auto now has a LinearOpMode as instance data which refers to the autonomous program that is using it. This enables us to use the isStopRequested() function which can only be accessed in Linear OpModes. If this works tmrw, we will add while loop breaking to OmniSlideDrive by adding individual setup methods for setting the LinearOpMode and startTime (rather than instance data defined in the constructor)